changes
I formatted a little bit more and moved some stuff around in our eda section. I then edited the descriptions for all of the relationship eda graphs. I changed the heatmap to delete all the na values and now it looks a lot better. Next steps would be to work on the model and description, evaluating the success of the model and what the next steps would be in the future